% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{echoAirGetMeta}
\alias{echoAirGetMeta}
\title{Downloads EPA ECHO Air Facility Metadata}
\usage{
echoAirGetMeta(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE}
}
\value{
returns a dataframe
}
\description{
Returns variable name and descriptions for parameters returned by \code{\link{echoAirGetFacilityInfo}}
}
\examples{
\dontrun{
## These examples require an internet connection to run

# returns a dataframe of
echoAirGetMeta()
}
}
