% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{echoGetEffluent}
\alias{echoGetEffluent}
\title{Downloads EPA ECHO DMR records}
\usage{
echoGetEffluent(...)
}
\arguments{
\item{...}{}
}
\value{
dataframe
}
\description{
Downloads EPA ECHO DMR records
}
\examples{
\dontrun{
## Retrieve single DMR for flow
echoGetEffluent(p_id = "tx0119407", parameter_code = "50050")
}
}
