% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoGetReports.R
\name{echoGetReports}
\alias{echoGetReports}
\title{Downloads self reported discharge and emissions data}
\usage{
echoGetReports(program, p_id, verbose = FALSE, ...)
}
\arguments{
\item{program}{Character, one of \code{program = 'caa'}, \code{program = 'cwa'}, or \code{program = 'nncr'}. \code{'caa'} retrieves facilities permitted under the Clean Air Act, \code{'cwa'} retrieves facilities permitted under the Clean Water Act, and \code{'nncr'} retrieves NPDES Noncompliance Report (NNCR) data for a facility or NPDES permit.}

\item{p_id}{Character string specify the identifier for the service. For \code{program = 'nncr'}, this is treated as the NPDES permit ID (\code{permits__npdes_id}). Required.}

\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE.}

\item{...}{Further arguments passed on as query parameters sent to EPA's ECHO API.}
}
\value{
Returns a dataframe
}
\description{
Downloads self reported discharge and emissions data
}
