% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{echoWaterGetMeta}
\alias{echoWaterGetMeta}
\title{Downloads EPA ECHO Water Facility Metadata}
\usage{
echoWaterGetMeta(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE}
}
\value{
returns a dataframe
}
\description{
Returns variable name and descriptions for parameters returned by \code{\link{echoWaterGetFacilityInfo}}
}
\examples{
\dontrun{
## These examples require an internet connection to run

# returns a dataframe of
echoWaterGetMeta()
}
}
