% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{echoWaterGetParams}
\alias{echoWaterGetParams}
\title{Search parameter codes for Clean Water Act permits on EPA ECHO}
\usage{
echoWaterGetParams(term = NULL, code = NULL, verbose = FALSE)
}
\arguments{
\item{term}{Character string specifying the parameter search term. Partial or complete search phrase or word.}

\item{code}{Character string specifying the parameter search code value.}

\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE}
}
\value{
Returns a dataframe.
}
\description{
Returns a dataframe of parameter codes and descriptions.
}
\examples{
\dontrun{
## Retrieve parameter codes for dissolved oxygen
echoWaterGetParams(term = "Oxygen, dissolved")

echoWaterGetParams(code = "00300")
}
}
