% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transfer_steps.R
\name{get_transfer_steps}
\alias{get_transfer_steps}
\title{Write the transfer steps for compounds}
\usage{
get_transfer_steps(
  standardized_daugther,
  standarized_mother,
  .echo_drop_nL = 25,
  missing_cmpds = "drop",
  conc_error = "drop"
)
}
\arguments{
\item{standardized_daugther}{The daughter plate layout, standardized for echo, as done by standardize_layout(which_plate = "daughter")}

\item{standarized_mother}{The mother plate layout, standardized for echo, as done by standardize_layout(which_plate = "mother")}

\item{.echo_drop_nL}{The droplet volume for the echo to be used, in nL. Defaults to 25.}

\item{missing_cmpds}{How to handle compounds present in the daughter plate, but not the mother. Options are "stop", which halts execution and errors, and "drop", which removes all of these wells from the daughter layout. Defaults to "stop".}

\item{conc_error}{How to handle compounds present in the daughter plate at higher concentration than in the mother (impossible to achieve). Options are "stop", which halts execution and errors, and "drop", which removes all of these wells from the daughter layout, "scale_down", which takes all occurances of that compound in the daughter plate and reduces their concentration such that relative dilutions are unchanged, but scaled down so that the highest concentration is the same as the concentration in the mother plate, and "make_max", which replaces these concentrations with the maximum achievable concentration, which is the concentration in the mother plate. Defaults to "stop".}
}
\value{
A tibble containing all transfer steps for all compounds in the daughter plate. Transer steps for dilutions are appended in write_echo_instructions().
}
\description{
This function handles inconsistencies and errors in the daughter layout. The dilution transfer steps and finalization of formatting for reading by the echo are handled by write_echo_instructions()
}
