% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missing_compounds.R
\name{handle_missing_compounds}
\alias{handle_missing_compounds}
\title{Resolve cases when a compound is not present in the mother}
\usage{
handle_missing_compounds(mother, daughter, missing_cmpds = "stop")
}
\arguments{
\item{mother}{mother plate layout, standardized for echo functions with standardize_layout()}

\item{daughter}{daughter plate layout, standardized for echo functions with standardize_layout()}

\item{missing_cmpds}{how to handle missing compounds. "stop" will produce an error, halting execution. "drop" will remove the missing compounds from the daughter layout.}
}
\value{
if there are no missing compounds, returns the daughter layout unchanged. If there are missing compounds, returns the daughterlayout with the missing compounds removed if missing_cmps is set to "drop", or an error, if missing_cmods is set to "stop"
}
\description{
Allow users to choose how to resolve cases when a compound in the daughter plate is not present in the mother plate, and therefore cannot be transferred. Compounds are mapped between the daughter and mother plates by name, so compounds must have exactly the the same name in the mother and daughter plates for the trasnfer instructions to be written (case sensitive).
}
