% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_varied_mother_concentrations.R
\name{handle_varied_mother_concentrations}
\alias{handle_varied_mother_concentrations}
\title{Handle varied compound concentrations in the mother plate}
\usage{
handle_varied_mother_concentrations(mother, keep_which = "max_conc")
}
\arguments{
\item{mother}{mother plate layout, standardized for echo functions with standardize_layout()}

\item{keep_which}{how to handle compounds that appear in the mother at multiple concentrations."stop" will produce an error, halting execution. "max_conc" will keep only the well(s) with the highest concentration. "most_wells" will keep the concentration in the greatest number of wells, removing all wells which do not have that compound at that concentration.}
}
\value{
the mother layout, with all compounds occuring in only a single concentration. The wells containing that compound at a different concentration have been removed.
}
\description{
This version of the code requires that each compound in the mother occurs at only one concentration. This function checks if any compounds are in the mother plateat muliple concentartions, and resolves this by either halting execution in an error (used as a stopping point for user-defined action), using only the well(s) with the highest concentration, or using only the wells which contain the most represented concentration.
}
