% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dilutions_plate.R
\name{make_dilutions_plate}
\alias{make_dilutions_plate}
\title{Calculate dilution transfers}
\usage{
make_dilutions_plate(
  transfers,
  standardized_mother,
  .echo_drop_nL = 25,
  .dil_name = "DMSO"
)
}
\arguments{
\item{transfers}{a tibble containging the transfer steps, as calculated by get_transfer_steps()}

\item{standardized_mother}{The mother plate layout, standardized for echo, as done by standardize_layout(which_plate = "mother")}

\item{.echo_drop_nL}{The droplet volume for the echo to be used, in nL. Defaults to 25.}

\item{.dil_name}{The name, as listed in the mother plate, of the substance to be used for dilutions. Defaults to "DMSO"}
}
\value{
all transfers needed to dilute compounds to the requested volume in the daughter plate.
}
\description{
Calculate dilution transfers
}
