% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_conc_error.R
\name{repair_conc_error}
\alias{repair_conc_error}
\title{Repair daughter plates that request impossible concentrations}
\usage{
repair_conc_error(transfers, conc_error = "stop", .echo_drop_nL = 25)
}
\arguments{
\item{transfers}{a tibble containging the transfer steps, as calculated by get_transfer_steps()}

\item{conc_error}{action to take to resolve the request. "stop" will throw an error, and halt execution. "drop" removes these wells from the daughter palte. "scale_down" reduces all concentrations of that compound in the daugther by an equal factor, such that the tested concentrations remain the same relative to one another, but are all at or below the concentration present in the mother. If "make_max", overwrites the concentration of that compound in the daughter to the concentration in the mother in only the wells in which the requested daughter concentration is higher than the concentration of that compound in the mother.}

\item{.echo_drop_nL}{the tarnsfer drop volume for the echo to be used. Defaults to 25 nL.}
}
\value{
the transfers tibble, with all concentrations in the daughter at or below the concentration in the mother.
}
\description{
If the concentration of a compound in the daughter plate is higher than its concentration in the mother plate, the transfer cannot be completed. This function catches and handles that scenario.
}
