% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resp_check}
\alias{resp_check}
\title{Check responses}
\usage{
resp_check(response)
}
\arguments{
\item{response}{response a [httr::GET()] request result returned from the API}
}
\value{
nothing if check is passed, or an informative error if not passed.
}
\description{
Checks for valid server response and passes silently or produces a hopefully useful error.
}
\keyword{internal}
