% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_echo_instructions.R
\name{write_echo_instructions}
\alias{write_echo_instructions}
\title{Write echo instructions}
\usage{
write_echo_instructions(
  trans_data,
  standardized_mother,
  .dil_name = "DMSO",
  .max_pull_nL = 40000
)
}
\arguments{
\item{trans_data}{A tibble containing the calculated transfer steps, as created by get_transfer_steps()}

\item{standardized_mother}{The conditions}

\item{.dil_name}{The name of the compound in the mother used for dilutions. Defaults to "DMSO".}

\item{.max_pull_nL}{The maximum volume (in nL) that can be pulled from any one well in the mother plate for dilutions.}
}
\value{
A tibble which can be saved as a csv and uploaded directly to the Echo Plate Reformat software.
}
\description{
Given the transfer steps calculated with get_transfer_steps(), write the full echo instructions file. This includes adding the dilution steps, standardizing column names, and adding additional columns used by the Echo Reformat plate software. Includes a check that none of the wells in the mother used for dilutions will be used to transfer a greater volume than they contain. This volume, given in nL, defaults to 40000 -- equivalent to 40 uL.
}
