% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dilutions.R
\name{calculate_dilutions}
\alias{calculate_dilutions}
\title{Calculate dilutions between mother and daughter plates}
\usage{
calculate_dilutions(cmpd_joined, .echo_drop_nL = 25)
}
\arguments{
\item{cmpd_joined}{a tibble, resulting from joining a standardized mother and standardized daughter layout.}

\item{.echo_drop_nL}{the droplet size of the Echo to be used, in nL. Defaults to 25.}
}
\value{
the input tibble, with additional columns describing the volume to be transferred from the mother, and the extent to which rounding the input concentration was necessary, given the droplet size of the echo used.
}
\description{
A helper funciton used inside of calculate_transfers() to calculate the volume of compound to transfer from mother to daughter wells.
}
