% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{downloadDMRs}
\alias{downloadDMRs}
\title{Title}
\usage{
downloadDMRs(df, nColumn, pBar = TRUE, ...)
}
\arguments{
\item{df}{dataframe with column of id numbers}

\item{nColumn}{numeric, indicates the column with p_id permit numbers}

\item{pBar}{logical, display a progress bar? Defaults to TRUE}

\item{...}{additional arguments passed to echoGetEffluentSummary}
}
\value{
dataframe df, with a column containing the discharge monitoring reports downloaded with echoGetEffluentSummary
}
\description{
Title
}
