% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{echoAirGetFacilityInfo}
\alias{echoAirGetFacilityInfo}
\title{Downloads permitted air discharger information from EPA ECHO}
\usage{
echoAirGetFacilityInfo(..., output)
}
\arguments{
\item{...}{see \url{https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_air_rest_services_get_facility_info} for a complete list of parameter options. Examples provided below.}

\item{output}{character string specifying output format. One of "JSON" or "GEOJSON"}
}
\value{
dataframe or geojson suitable for plotting
}
\description{
Downloads permitted air discharger information from EPA ECHO
}
\examples{
\dontrun{
## Not run:
## Retrieve table of facilities by bounding box
echoAirGetFacilityInfo(xmin = "-96.407563",
ymin = "30.554395",
xmax = "-96.25947",
ymax = "30.751984",
output = "JSON")

## Retrieve a geojson by bounding box
spatialdata <- echoAirGetFacilityInfo(xmin = "-96.407563",
ymin = "30.554395",
xmax = "-96.25947",
ymax = "30.751984",
output = "GEOJSON")

leaflet() \%>\%
    addTiles() \%>\%
    addGeoJSON(geojson = spatialdata)
}
}
