% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{echoAirGetFacilityInfo}
\alias{echoAirGetFacilityInfo}
\title{Downloads permitted air discharger information from EPA ECHO}
\usage{
echoAirGetFacilityInfo(output = "JSON", verbose = FALSE, ...)
}
\arguments{
\item{output}{character string specifying output format. \code{output = 'JSON'} or \code{output = 'GEOJSON'}}

\item{verbose}{Logical, indicating whether to provide prcessing and retrieval messages. Defaults to FALSE}

\item{...}{see \url{https://echo.epa.gov/tools/web-services/facility-search-water#!/Facility_Information/get_air_rest_services_get_facility_info} for a complete list of parameter options. Examples provided below.}
}
\value{
dataframe or geojson suitable for plotting
}
\description{
Downloads permitted air discharger information from EPA ECHO
}
\examples{
\dontrun{
## Not run:
## Retrieve table of facilities by bounding box
echoAirGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'JSON')

## Retrieve a geojson by bounding box
spatialdata <- echoAirGetFacilityInfo(xmin = '-96.407563',
ymin = '30.554395',
xmax = '-96.25947',
ymax = '30.751984',
output = 'GEOJSON')

leaflet() \%>\%
    addTiles() \%>\%
    addGeoJSON(geojson = spatialdata)
}
}
