% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{echoGetCAAPR}
\alias{echoGetCAAPR}
\title{Download EPA ECHO emissions inventory report data}
\usage{
echoGetCAAPR(...)
}
\arguments{
\item{...}{}
}
\value{
dataframe
}
\description{
Download EPA ECHO emissions inventory report data
}
\examples{
\dontrun{
echoGetCAAPR(p_id = "110000350174")
}

}
