% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air.R
\name{echoGetCAAPR}
\alias{echoGetCAAPR}
\title{Download EPA ECHO emissions inventory report data}
\usage{
echoGetCAAPR(p_id, verbose = FALSE, ...)
}
\arguments{
\item{p_id}{character string specify the identifier for the service. Required.}

\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE}

\item{...}{Additional arguments}
}
\value{
dataframe
}
\description{
Download EPA ECHO emissions inventory report data
}
\examples{
\dontrun{
## This example requires an internet connection to run

echoGetCAAPR(p_id = '110000350174')
}

}
