% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoGetReports.R
\name{echoGetReports}
\alias{echoGetReports}
\title{Downloads self reported discharge and emissions data}
\usage{
echoGetReports(program, p_id, verbose = FALSE, ...)
}
\arguments{
\item{program}{Character, either \code{program = 'caa'} or \code{program = 'cwa'}. \code{'caa'} retrieves facilities permitted under the Clean Air Act, \code{'cwa'} retreives facilites permitted under the Clean Water Act.}

\item{p_id}{Character string specify the identifier for the service. Required.}

\item{verbose}{Logical, indicating whether to provide processing and retrieval messages. Defaults to FALSE.}

\item{...}{Further arguments passed on as query parameters sent to EPA's ECHO API.}
}
\value{
Returns a dataframe
}
\description{
Downloads self reported discharge and emissions data
}
